+++ src/main/java/net/spoiledz/access/ServerWorldAccess.java
@@ -1,0 +1,9 @@
+package net.spoiledz.access;
+
+import net.spoiledz.util.SpoiledUtil.FoodBlockMap;
+
+public interface ServerWorldAccess {
+
+    public FoodBlockMap getFoodBlockMap();
+
+}