+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### Added:
-- Added spoilage season count config option
+- Added food block spoilage
 ### Fixed:
 - 
 ### Changed: